requirement.txt
aiofiles==23.2.1
aiosmtplib==2.0.2
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
asyncmy==0.2.9
bcrypt==4.1.2
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.0
fastapi-mail==1.4.1
h11==0.14.0
httptools==0.6.1
idna==3.6
iso8601==1.1.0
Jinja2==3.1.3
MarkupSafe==2.1.5
passlib==1.7.4
pillow==10.2.0
pycparser==2.21
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyJWT==2.8.0
pypika-tortoise==0.1.6
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
sniffio==1.3.1
starlette==0.36.3
tortoise-orm==0.20.0
typing_extensions==4.10.0
uvicorn==0.28.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
